# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 鲸云AI教育平台 API
  version: 0.1.0
  description: |
    W0-2 契约 · MVP 口径(对应《MVP 裁剪手册》)。统一响应包 {code,message,data};
    分页 {items,total};鉴权 Bearer JWT,payload 含 {uid,orgId,role}。
    角色约束写在各接口 description 的 [角色] 标注中。
servers: [{ url: /api/v1 }]
security: [{ bearer: [] }]
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    page: { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
    size: { name: size, in: query, schema: { type: integer, default: 20, maximum: 50 } }
    keyword: { name: keyword, in: query, schema: { type: string } }
    idPath: { name: id, in: path, required: true, schema: { type: integer } }
  responses:
    Err:
      description: 业务/鉴权错误(401 未登录 403 越权 404 不存在或跨租户)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrResp' } } }
  schemas:
    ErrResp:
      type: object
      required: [code, message]
      properties: { code: { type: integer }, message: { type: string }, detail: {} }
    OkVoid:
      type: object
      required: [code, message, data]
      properties: { code: { type: integer, enum: [0] }, message: { type: string }, data: { type: 'null' } }
    # ---- 枚举 ----
    Role: { type: string, enum: [admin, teacher, student] }
    UserStatus: { type: string, enum: [active, disabled, pending] }
    ClassType: { type: string, enum: [group, one_on_one, one_on_three] }
...
    GraphType: { type: string, enum: [curriculum_knowledge, problem_solving_ability, problem_solving_strategy] }
...
    Me:
      type: object
      required: [id, orgId, role, name, orgName, orgSettings]
      properties:
        id: { type: integer }
        orgId: { type: integer }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string }
        orgName: { type: string }
        orgSettings:
          type: object
          required: [ai, studentHours]
          properties:
            ai:
              type: object
              required: [qaGuideOnly, preGrading]
              properties: { qaGuideOnly: { type: boolean }, preGrading: { type: boolean }, classCompanion: { type: boolean }, diagnosis: { type: boolean } }
            studentHours:
              type: object
              required: [start, end]
              properties: { start: { type: string }, end: { type: string } }
    Teacher:
      type: object
      required: [id, name, teacherNo, phone, stage, subject, status, courseCount, questionCount, resourceCount]
      properties:
        id: { type: integer }
        name: { type: string }
        teacherNo: { type: string }
        phone: { type: string }
        stage: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/UserStatus' }
        courseCount: { type: integer }
        questionCount: { type: integer }
        resourceCount: { type: integer }
    TeacherInput:
      type: object
      required: [name, phone, stage, subject]
      properties:
        name: { type: string, maxLength: 32 }
        phone: { type: string }
        teacherNo: { type: string, description: 不传则自动生成 }
        stage: { type: string }
        subject: { type: string }
    Student:
      type: object
      required: [id, name, studentNo, parentPhone, grade, status, courses, weekStudySec]
      properties:
        id: { type: integer }
        name: { type: string }
        studentNo: { type: string }
        parentPhone: { type: string }
        grade: { type: string }
        status: { $ref: '#/components/schemas/UserStatus' }
        courses:
          type: array
          items:
            type: object
            required: [id, name, classType]
            properties:
              id: { type: integer }
              name: { type: string }
              classType: { $ref: '#/components/schemas/ClassType' }
        weekStudySec: { type: integer }
    StudentInput:
      type: object
      required: [name, parentPhone, grade]
      properties:
        name: { type: string, maxLength: 32 }
        parentPhone: { type: string }
        studentNo: { type: string }
        grade: { type: string }
        courseIds: { type: array, items: { type: integer } }
    Course:
      type: object
      required: [id, name, classType, subject, stage, teacherId, teacherName, totalLessons, currentLesson, studentCount, status, nextLessonAt, attendanceRate, homeworkRate]
      properties:
        id: { type: integer }
        name: { type: string }
        classType: { $ref: '#/components/schemas/ClassType' }
        subject: { type: string }
        stage: { type: string }
        teacherId: { type: integer }
        teacherName: { type: string }
        totalLessons: { type: integer }
        currentLesson: { type: integer }
        studentCount: { type: integer }
        status: { type: string }
        nextLessonAt: { oneOf: [{ type: 'null' }, { type: string, format: date-time }] }
        attendanceRate: { oneOf: [{ type: 'null' }, { type: number }] }
        homeworkRate: { oneOf: [{ type: 'null' }, { type: number }] }
    CourseInput:
      type: object
      required: [name, classType, subject, stage, teacherId, totalLessons]
      properties:
        name: { type: string }
        classType: { $ref: '#/components/schemas/ClassType' }
        subject: { type: string }
        stage: { type: string }
        teacherId: { type: integer }
        totalLessons: { type: integer, minimum: 1 }
        studentIds: { type: array, items: { type: integer } }
...
    MasteryItem:
      type: object
      required: [nodeId, nodeName, graphType, mastery, sampleCount]
      properties:
        nodeId: { type: integer }
        nodeName: { type: string }
        graphType: { $ref: '#/components/schemas/GraphType' }
        mastery: { type: integer, minimum: 0, maximum: 100 }
        sampleCount: { type: integer }
    AiUsageSummary:
      type: object
      required: [period, totalTokens, totalCost, monthlyLimit, usedPercent, avgCostPerLesson]
      properties:
        period: { type: string }
        totalTokens: { type: integer }
        totalCost: { type: number }
        monthlyLimit: { type: number }
        usedPercent: { type: number }
        avgCostPerLesson: { oneOf: [{ type: 'null' }, { type: number }] }
    AiUsageBreakdown:
      type: object
      required: [key, label, tokens, cost, percent]
      properties:
        key: { type: string }
        label: { type: string }
        tokens: { type: integer }
        cost: { type: number }
        percent: { type: number }
    # ---- AI 接口管理(admin)· 运行态 LLM 配置;与 src/dto.ts 一一对应 ----
    AiProviderConfig:
      type: object
      required: [baseUrl, model, apiKeyMasked, concurrency, source]
      properties:
        baseUrl: { type: string }
        model: { type: string }
        apiKeyMasked: { type: string, description: 脱敏后的 key,如 sk-****e86f3;未配置则空串(绝不回明文) }
        concurrency: { type: integer, description: 全局并发上限(同时在飞的 LLM 调用数) }
        source: { type: string, enum: [runtime, env], description: 当前生效来源:runtime=Redis 运行态 / env=兜底 }
    AiProviderConfigInput:
      type: object
      required: [baseUrl, model, concurrency]
      properties:
        baseUrl: { type: string }
        model: { type: string }
        apiKey: { type: string, description: 留空=保留现有,不覆盖 }
        concurrency: { type: integer }
    AiFeatureRoutes:
      type: object
      required: [qa, pre_grading, class_companion, diagnosis]
      properties:
        qa: { type: string, enum: [real, mock] }
        pre_grading: { type: string, enum: [real, mock] }
        class_companion: { type: string, enum: [real, mock] }
        diagnosis: { type: string, enum: [real, mock] }
    AiTestResult:
      type: object
      required: [ok, latencyMs, sample, error]
      properties:
        ok: { type: boolean }
        latencyMs: { type: integer }
        sample: { oneOf: [{ type: 'null' }, { type: string }], description: 成功时模型回的一小段文本 }
        error: { oneOf: [{ type: 'null' }, { type: string }], description: 失败原因 }
...
paths:
...
  /admin/teachers:
    get:
      summary: 教师列表 [admin]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}, {$ref: '#/components/parameters/keyword'}, {name: status, in: query, schema: {$ref: '#/components/schemas/UserStatus'}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Teacher'}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    post:
      summary: 创建教师 [admin]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TeacherInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Teacher'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/teachers/{id}:
    put:
      summary: 编辑教师 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TeacherInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
    delete:
      summary: 停用教师(软删) [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/teachers/{id}/reset-password:
    post:
      summary: 重置教师密码 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [password], properties: {password: {type: string}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/teachers/{id}/enable:
    post:
      summary: 启用教师 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/students:
    get:
      summary: 学生列表 [admin]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}, {$ref: '#/components/parameters/keyword'}, {name: status, in: query, schema: {$ref: '#/components/schemas/UserStatus'}}, {name: courseId, in: query, schema: {type: integer}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Student'}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    post:
      summary: 创建学生(向家长手机发登录码) [admin]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StudentInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Student'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/students/{id}:
    put:
      summary: 编辑学生 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StudentInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
    delete:
      summary: 停用学生(软删) [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/students/{id}/profile:
    get:
      summary: 学生档案(课程+设备+学情摘要) [admin/teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [student, mastery, wrongOpenCount], properties: {student: {$ref: '#/components/schemas/Student'}, mastery: {type: array, items: {$ref: '#/components/schemas/MasteryItem'}}, wrongOpenCount: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/students/{id}/reset-password:
    post:
      summary: 重置学生密码 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [password], properties: {password: {type: string}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/students/{id}/enable:
    post:
      summary: 启用学生 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/courses:
    get:
      summary: 课程列表 [admin]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}, {$ref: '#/components/parameters/keyword'}, {name: classType, in: query, schema: {$ref: '#/components/schemas/ClassType'}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Course'}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    post:
      summary: 创建课程(自动生成 totalLessons 个空讲次) [admin]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CourseInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Course'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/courses/{id}:
    put:
      summary: 编辑课程 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CourseInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/courses/{id}/roster:
    get:
      summary: 课程名单(到课/作业概览) [admin/teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {type: object, required: [studentId, name, attendance, homeworkAvg, status], properties: {studentId: {type: integer}, name: {type: string}, attendance: {type: string}, homeworkAvg: {oneOf: [{type: 'null'}, {type: number}]}, status: {type: string}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/courses/{id}/students:
    post:
      summary: 课程添加学生 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [studentIds], properties: {studentIds: {type: array, items: {type: integer}}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/courses/{id}/students/{studentId}:
    delete:
      summary: 课程移除学生 [admin]
      parameters: [{$ref: '#/components/parameters/idPath'}, {name: studentId, in: path, required: true, schema: {type: integer}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/dashboard:
    get:
      summary: 管理员总览 [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [teacherCount, studentCount, weekAttendanceRate, monthAiCost, todayLessonCount, recentEvents], properties: {teacherCount: {type: integer}, studentCount: {type: integer}, weekAttendanceRate: {oneOf: [{type: 'null'}, {type: number}]}, monthAiCost: {type: number}, todayLessonCount: {type: integer}, recentEvents: {type: array, items: {type: object, required: [text, time], properties: {text: {type: string}, time: {type: string, format: date-time}}}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai-usage/summary:
    get:
      summary: 本月 AI 用量摘要 [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/AiUsageSummary'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai-usage/daily:
    get:
      summary: 近 N 日用量曲线 [admin]
      parameters: [{name: days, in: query, schema: {type: integer, default: 14, maximum: 31}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {type: object, required: [date, tokens, cost], properties: {date: {type: string}, tokens: {type: integer}, cost: {type: number}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai-usage/breakdown:
    get:
      summary: 按功能拆分(本月) [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {$ref: '#/components/schemas/AiUsageBreakdown'}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai-quota:
    get:
      summary: 读取额度与告警设置 [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [monthlyLimit, alertThreshold, overPolicy], properties: {monthlyLimit: {type: number}, alertThreshold: {type: integer}, overPolicy: {type: string}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    put:
      summary: 修改额度与告警 [admin]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [monthlyLimit, alertThreshold, overPolicy], properties: {monthlyLimit: {type: number}, alertThreshold: {type: integer, minimum: 50, maximum: 95}, overPolicy: {type: string, enum: [disable_qa, pause_all, record_only]}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai/config:
    get:
      summary: 读取运行态 LLM 供应商配置(key 脱敏) [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/AiProviderConfig'}}}}}}
        default: {$ref: '#/components/responses/Err'}
    put:
      summary: 写入运行态 LLM 供应商配置(apiKey 留空=不覆盖,运行态生效不重启) [admin]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiProviderConfigInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai/routes:
    get:
      summary: 读取逐功能真假路由 [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/AiFeatureRoutes'}}}}}}
        default: {$ref: '#/components/responses/Err'}
    put:
      summary: 设置逐功能真假路由(逐功能切真实↔mock) [admin]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiFeatureRoutes'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/ai/test:
    post:
      summary: 测试 LLM 连接(可选指定功能) [admin]
      requestBody: {content: {application/json: {schema: {type: object, properties: {feature: {type: string}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/AiTestResult'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/settings:
    get:
      summary: 平台设置 [admin]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Me'}}}}}}
        default: {$ref: '#/components/responses/Err'}
    put:
      summary: 修改设置(引导模式/使用时段/AI 功能开关) [admin]
      requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {qaGuideOnly: {type: boolean}, preGrading: {type: boolean}, classCompanion: {type: boolean}, diagnosis: {type: boolean}, studentHours: {type: object, required: [start, end], properties: {start: {type: string}, end: {type: string}}}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /admin/audit-logs:
    get:
      summary: 审计日志 [admin]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {type: object, required: [actorName, action, createdAt], properties: {actorName: {type: string}, action: {type: string}, targetType: {oneOf: [{type: 'null'}, {type: string}]}, createdAt: {type: string, format: date-time}}}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}

  # ================= 知识图谱(只读) =================
...